Factor out y coordinates of control chart plot points
so we can start writing control chart rules on these y coordinates
generically. Not every chart works on the means or the like. Some of
them work on the ranges, etc. The rules of control charts, as in
Shewhart's or Western Electric rules operate just on the y of a point
generically.